@@ -1,22 +1,14 @@
-﻿using System;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.UI;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using SwitchBook.Data;
-using SwitchBook.Models;
+﻿using Microsoft.AspNetCore.Hosting;
+using SwitchBook.Areas.Identity;
 
-[assembly: HostingStartup(typeof(SwitchBook.Areas.Identity.IdentityHostingStartup))]
-namespace SwitchBook.Areas.Identity
+[assembly: HostingStartup(typeof(IdentityHostingStartup))]
+
+namespace SwitchBook.Areas.Identity;
+
+public class IdentityHostingStartup : IHostingStartup
 {
-    public class IdentityHostingStartup : IHostingStartup
+    public void Configure(IWebHostBuilder builder)
     {
-        public void Configure(IWebHostBuilder builder)
-        {
-            builder.ConfigureServices((context, services) => {
-            });
-        }
+        builder.ConfigureServices((context, services) => { });
     }
 }
