@@ -5,30 +5,26 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
 using SwitchBook.Models;
 
-namespace SwitchBook.Areas.Identity.Pages.Account.Manage
+namespace SwitchBook.Areas.Identity.Pages.Account.Manage;
+
+public class PersonalDataModel : PageModel
 {
-    public class PersonalDataModel : PageModel
+    private readonly ILogger<PersonalDataModel> _logger;
+    private readonly UserManager<User> _userManager;
+
+    public PersonalDataModel(
+        UserManager<User> userManager,
+        ILogger<PersonalDataModel> logger)
     {
-        private readonly UserManager<User> _userManager;
-        private readonly ILogger<PersonalDataModel> _logger;
+        _userManager = userManager;
+        _logger = logger;
+    }
 
-        public PersonalDataModel(
-            UserManager<User> userManager,
-            ILogger<PersonalDataModel> logger)
-        {
-            _userManager = userManager;
-            _logger = logger;
-        }
+    public async Task<IActionResult> OnGet()
+    {
+        var user = await _userManager.GetUserAsync(User);
+        if (user == null) return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
 
-        public async Task<IActionResult> OnGet()
-        {
-            var user = await _userManager.GetUserAsync(User);
-            if (user == null)
-            {
-                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
-            }
-
-            return Page();
-        }
+        return Page();
     }
 }
