@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.IO;
 using System.Linq;
@@ -18,142 +17,142 @@ using Microsoft.Extensions.Logging;
 using SwitchBook.Data;
 using SwitchBook.Models;
 
-namespace SwitchBook.Areas.Identity.Pages.Account
+namespace SwitchBook.Areas.Identity.Pages.Account;
+
+[AllowAnonymous]
+public class RegisterModel : PageModel
 {
-    [AllowAnonymous]
-    public class RegisterModel : PageModel
+    private readonly ApplicationDbContext _db;
+    private readonly IEmailSender _emailSender;
+    private readonly ILogger<RegisterModel> _logger;
+    private readonly SignInManager<User> _signInManager;
+    private readonly UserManager<User> _userManager;
+
+    public RegisterModel(
+        UserManager<User> userManager,
+        SignInManager<User> signInManager,
+        ILogger<RegisterModel> logger,
+        IEmailSender emailSender, ApplicationDbContext context)
     {
-        private readonly SignInManager<User> _signInManager;
-        private readonly UserManager<User> _userManager;
-        private readonly ILogger<RegisterModel> _logger;
-        private readonly IEmailSender _emailSender;
-        private readonly ApplicationDbContext _db;
-        public RegisterModel(
-            UserManager<User> userManager,
-            SignInManager<User> signInManager,
-            ILogger<RegisterModel> logger,
-            IEmailSender emailSender, ApplicationDbContext context)
+        _userManager = userManager;
+        _signInManager = signInManager;
+        _logger = logger;
+        _emailSender = emailSender;
+        _db = context;
+    }
+
+    [BindProperty] public InputModel Input { get; set; }
+
+    public string ReturnUrl { get; set; }
+
+    public IList<AuthenticationScheme> ExternalLogins { get; set; }
+
+    public async Task OnGetAsync(string returnUrl = null)
+    {
+        ReturnUrl = returnUrl;
+        ExternalLogins = (await _signInManager.GetExternalAuthenticationSchemesAsync()).ToList();
+    }
+
+    public async Task<IActionResult> OnPostAsync(string returnUrl = null)
+    {
+        returnUrl ??= Url.Content("~/");
+        ExternalLogins = (await _signInManager.GetExternalAuthenticationSchemesAsync()).ToList();
+        if (ModelState.IsValid)
         {
-            _userManager = userManager;
-            _signInManager = signInManager;
-            _logger = logger;
-            _emailSender = emailSender;
-            _db = context;
-        }
-
-        [BindProperty]
-        public InputModel Input { get; set; }
-
-        public string ReturnUrl { get; set; }
-
-        public IList<AuthenticationScheme> ExternalLogins { get; set; }
-
-        public class InputModel
-        {
-            [Required]
-            [EmailAddress]
-            [Display(Name = "Email")]
-            public string Email { get; set; }
-
-            [Required]
-            [StringLength(100, ErrorMessage = "The {0} must be at least {2} and at max {1} characters long.", MinimumLength = 6)]
-            [DataType(DataType.Password)]
-            [Display(Name = "Password")]
-            public string Password { get; set; }
-
-            [DataType(DataType.Password)]
-            [Display(Name = "Confirm password")]
-            [Compare("Password", ErrorMessage = "The password and confirmation password do not match.")]
-            public string ConfirmPassword { get; set; }
-            [DataType(DataType.Upload)]
-            [Display(Name = "Avatar")]
-            public IFormFile Avatar { get; set; }
-
-            [Display(Name = "Region")]
-
-            public string Region { get; set; }
-            [Display(Name = "City")]
-            public string City { get; set; }
-            [Display(Name = "Street")]
-            public string Street { get; set; }
-            [Display(Name = "PostalCode")]
-            public string PostalCode { get; set; }
-            [Display(Name = "PhoneNumber")]
-            [DataType(DataType.PhoneNumber)]
-            public string PhoneNumber { get; set; }
-        }
-
-        public async Task OnGetAsync(string returnUrl = null)
-        {
-            ReturnUrl = returnUrl;
-            ExternalLogins = (await _signInManager.GetExternalAuthenticationSchemesAsync()).ToList();
-        }
-
-        public async Task<IActionResult> OnPostAsync(string returnUrl = null)
-        {
-            returnUrl ??= Url.Content("~/");
-            ExternalLogins = (await _signInManager.GetExternalAuthenticationSchemesAsync()).ToList();
-            if (ModelState.IsValid)
+            var user = new User
             {
-                var user = new User { UserName = Input.Email, Email = Input.Email};
-                if (Input.Avatar != null)
+                UserName = Input.Email,
+                Email = Input.Email
+            };
+            if (Input.Avatar != null)
+            {
+                byte[] imageData = null;
+                // считываем переданный файл в массив байтов
+                using (var binaryReader = new BinaryReader(Input.Avatar.OpenReadStream()))
                 {
-                    byte[] imageData = null;
-                    // считываем переданный файл в массив байтов
-                    using (var binaryReader = new BinaryReader(Input.Avatar.OpenReadStream()))
-                    {
-                        imageData = binaryReader.ReadBytes((int)Input.Avatar.Length);
-                    }
-                    // установка массива байтов
-                    user.Avatar = imageData;
+                    imageData = binaryReader.ReadBytes((int)Input.Avatar.Length);
                 }
-                
-                Address address = new Address()
-                {
-                    Region = Input.Region,
-                    City = Input.City,
-                    PostalCode = Input.PostalCode,
-                    PhoneNumber = Input.PhoneNumber,
-                    Street = Input.Street
-                };
-                await _db.AddAsync(address);
-                await _db.SaveChangesAsync();
-                //var id = _db.Address.Find(address).Id;
-                user.AddressId = address.Id;
-                var result = await _userManager.CreateAsync(user, Input.Password);
-                if (result.Succeeded)
-                {
-                    _logger.LogInformation("User created a new account with password.");
 
-                    var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
-                    code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
-                    var callbackUrl = Url.Page(
-                        "/Account/ConfirmEmail",
-                        pageHandler: null,
-                        values: new { area = "Identity", userId = user.Id, code = code, returnUrl = returnUrl },
-                        protocol: Request.Scheme);
-
-                    await _emailSender.SendEmailAsync(Input.Email, "Confirm your email",
-                        $"Please confirm your account by <a href='{HtmlEncoder.Default.Encode(callbackUrl)}'>clicking here</a>.");
-
-                    if (_userManager.Options.SignIn.RequireConfirmedAccount)
-                    {
-                        return RedirectToPage("RegisterConfirmation", new { email = Input.Email, returnUrl = returnUrl });
-                    }
-                    else
-                    {
-                        await _signInManager.SignInAsync(user, isPersistent: false);
-                        return LocalRedirect(returnUrl);
-                    }
-                }
-                foreach (var error in result.Errors)
-                {
-                    ModelState.AddModelError(string.Empty, error.Description);
-                }
+                // установка массива байтов
+                user.Avatar = imageData;
             }
 
-            // If we got this far, something failed, redisplay form
-            return Page();
+            var address = new Address
+            {
+                Region = Input.Region,
+                City = Input.City,
+                PostalCode = Input.PostalCode,
+                PhoneNumber = Input.PhoneNumber,
+                Street = Input.Street,
+            };
+            await _db.AddAsync(address);
+            await _db.SaveChangesAsync();
+            user.AddressId = address.Id;
+            var result = await _userManager.CreateAsync(user, Input.Password);
+            if (result.Succeeded)
+            {
+                _logger.LogInformation("User created a new account with password.");
+
+                var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
+                code = WebEncoders.Base64UrlEncode(Encoding.UTF8.GetBytes(code));
+                var callbackUrl = Url.Page(
+                    "/Account/ConfirmEmail",
+                    null,
+                    new { area = "Identity", userId = user.Id, code, returnUrl },
+                    Request.Scheme);
+
+                await _emailSender.SendEmailAsync(Input.Email, "Confirm your email",
+                    $"Please confirm your account by <a href='{HtmlEncoder.Default.Encode(callbackUrl)}'>clicking here</a>.");
+
+                if (_userManager.Options.SignIn.RequireConfirmedAccount)
+                {
+                    return RedirectToPage("RegisterConfirmation", new { email = Input.Email, returnUrl });
+                }
+
+                await _signInManager.SignInAsync(user, false);
+                return LocalRedirect(returnUrl);
+            }
+
+            foreach (var error in result.Errors) ModelState.AddModelError(string.Empty, error.Description);
         }
+
+        // If we got this far, something failed, redisplay form
+        return Page();
+    }
+
+    public class InputModel
+    {
+        [Required]
+        [EmailAddress]
+        [Display(Name = "Email")]
+        public string Email { get; set; }
+
+        [Required]
+        [StringLength(100, ErrorMessage = "Довжина {0} має бути не менше {2} і не більше {1} символів.",
+            MinimumLength = 6)]
+        [DataType(DataType.Password)]
+        [Display(Name = "Password")]
+        public string Password { get; set; }
+
+        [DataType(DataType.Password)]
+        [Display(Name = "Confirm password")]
+        [Compare("Password", ErrorMessage = "Пароль і пароль підтвердження не збігаються.")]
+        public string ConfirmPassword { get; set; }
+
+        [DataType(DataType.Upload)]
+        [Display(Name = "Avatar")]
+        public IFormFile Avatar { get; set; }
+
+        [Display(Name = "Region")] public string Region { get; set; }
+
+        [Display(Name = "City")] public string City { get; set; }
+
+        [Display(Name = "Street")] public string Street { get; set; }
+
+        [Display(Name = "PostalCode")] public string PostalCode { get; set; }
+
+        [Display(Name = "PhoneNumber")]
+        [DataType(DataType.PhoneNumber)]
+        public string PhoneNumber { get; set; }
     }
 }
