@@ -1,21 +1,17 @@
 ﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Text;
 using SwitchBook.Models;
 
-namespace SwitchBook.Data
+namespace SwitchBook.Data;
+
+public class ApplicationDbContext : IdentityDbContext<User>
 {
-    public class ApplicationDbContext : IdentityDbContext<User>
+    public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
+        : base(options)
     {
-        public DbSet<Book> Books { get; set; }
-        public DbSet<Address> Address { get; set; }
-        public DbSet<Order> Orders { get; set; }
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
-            : base(options)
-        {
-        }
     }
+
+    public DbSet<Book> Books { get; set; }
+    public DbSet<Address> Address { get; set; }
+    public DbSet<Order> Orders { get; set; }
 }
