@@ -1,63 +1,51 @@
 ﻿@page
 @model IndexModel
 @{
-    ViewData["Title"] = "Profile";
+    ViewData["Title"] = "Обліковий запис";
     ViewData["ActivePage"] = ManageNavPages.Index;
 }
 
 <h4>@ViewData["Title"]</h4>
-<partial name="_StatusMessage" model="Model.StatusMessage" />
+<partial name="_StatusMessage" model="Model.StatusMessage"/>
 <div class="row">
     <div class="col-md-6">
 
         <form id="profile-form" method="post">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group">
-                <label asp-for="Input.UserName"></label>
+                <label asp-for="Input.UserName">Ім'я користувача</label>
                 <input asp-for="Input.UserName" class="form-control"/>
             </div>
-            @*<div class="form-group">
-                <label asp-for="Input.PhoneNumber"></label>
-                <input asp-for="Input.PhoneNumber" class="form-control"/>
-                <span asp-validation-for="Input.PhoneNumber" class="text-danger"></span>
-            </div>*@
-            
-            @*address*@
-            @*<div class="form-group">
-                <label asp-for="Input.Avatar"></label>
-                <input asp-for="Input.Avatar" class="form-control" />
-                <span asp-validation-for="Input.Avatar" class="text-danger"></span>
-            </div>*@
             <div class="form-group">
-                <label asp-for="Input.Region"></label>
-                <input asp-for="Input.Region" class="form-control" />
+                <label asp-for="Input.Region">Область</label>
+                <input asp-for="Input.Region" class="form-control"/>
                 <span asp-validation-for="Input.Region" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="Input.City"></label>
-                <input asp-for="Input.City" class="form-control" />
+                <label asp-for="Input.City">Місто</label>
+                <input asp-for="Input.City" class="form-control"/>
                 <span asp-validation-for="Input.City" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="Input.Street"></label>
-                <input asp-for="Input.Street" class="form-control" />
+                <label asp-for="Input.Street">Вулиця</label>
+                <input asp-for="Input.Street" class="form-control"/>
                 <span asp-validation-for="Input.Street" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="Input.PostalCode"></label>
-                <input asp-for="Input.PostalCode" class="form-control" />
+                <label asp-for="Input.PostalCode">Поштовий індекс</label>
+                <input asp-for="Input.PostalCode" class="form-control"/>
                 <span asp-validation-for="Input.PostalCode" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="Input.PhoneNumber"></label>
-                <input asp-for="Input.PhoneNumber" class="form-control" />
+                <label asp-for="Input.PhoneNumber">Номер телефону</label>
+                <input asp-for="Input.PhoneNumber" class="form-control"/>
                 <span asp-validation-for="Input.PhoneNumber" class="text-danger"></span>
             </div>
-            <button id="update-profile-button" type="submit" class="btn btn-primary">Save</button>
+            <button id="update-profile-button" type="submit" class="btn btn_book">Зберегти</button>
         </form>
     </div>
 </div>
 
 @section Scripts {
-    <partial name="_ValidationScriptsPartial" />
+    <partial name="_ValidationScriptsPartial"/>
 }
