@@ -9,14 +9,14 @@
     }
 }
 
-<h2>Manage your account</h2>
+<h2>Керування обліковим записом</h2>
 
 <div>
-    <h4>Change your account settings</h4>
-    <hr />
+    <h4>Зміна налаштувань облікового запису</h4>
+    <hr/>
     <div class="row">
         <div class="col-md-3">
-            <partial name="_ManageNav" />
+            <partial name="_ManageNav"/>
         </div>
         <div class="col-md-9">
             @RenderBody()
@@ -25,5 +25,5 @@
 </div>
 
 @section Scripts {
-    @RenderSection("Scripts", required: false)
+    @RenderSection("Scripts", false)
 }
