@@ -1,6 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace SwitchBook.Models
+﻿namespace SwitchBook.Models
 {
     public class Book
     {
@@ -8,6 +6,7 @@ namespace SwitchBook.Models
         public string Title { get; set; }
         public string Author { get; set; }
         public string OwnerId { get; set; }
+        public User Owner { get; set; }
         public string Description { get; set; }
 
         public byte[] Image { get; set; }
