@@ -9,30 +9,31 @@
 <h4>Книга</h4>
 
 @section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
+    @{ await Html.RenderPartialAsync("_ValidationScriptsPartial"); }
 }
+
 <form enctype="multipart/form-data" method="post" asp-action="Create">
     <div asp-validation-summary="ModelOnly" class="text-danger"></div>
     <div class="row book_details">
         <dl class="col">
             <div class="form-group">
                 <label asp-for="Title" class="control-label">Назва</label>
-                <input asp-for="Title" class="form-control" />
+                <input asp-for="Title" class="form-control"/>
                 <span asp-validation-for="Title" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <label asp-for="Author" class="control-label">Автор</label>
-                <input asp-for="Author" class="form-control" />
+                <input asp-for="Author" class="form-control"/>
                 <span asp-validation-for="Author" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <label asp-for="Description" class="control-label">Відгук</label>
-                <input asp-for="Description" class="form-control" />
+                <input asp-for="Description" class="form-control"/>
                 <span asp-validation-for="Description" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <label asp-for="Image" class="control-label">Обкладинка</label>
-                <input asp-for="Image" name="Image" type="file" class="form-control" style="height: min-content; padding: 3px;" />
+                <input asp-for="Image" name="Image" type="file" class="form-control" style="height: min-content; padding: 3px;"/>
                 <span asp-validation-for="Image" class="text-danger"></span>
             </div>
         </dl>
@@ -40,7 +41,7 @@
     <div class="book_buttons">
         <hr/>
         <form asp-action="Delete">
-            <input type="submit" value="Створити" class="btn btn_book" />
+            <input type="submit" value="Створити" class="btn btn_book"/>
             <a asp-action="Index" class="btn btn_book">Повернутися без збереження змін</a>
         </form>
     </div>
