@@ -1,30 +1,28 @@
-﻿@using Microsoft.AspNetCore.Identity
+﻿@using SwitchBook.Models
 @using Microsoft.AspNetCore.Mvc.TagHelpers
-@using SwitchBook.Models
-@using System.Security.Claims
 @inject SignInManager<User> SignInManager
 @inject UserManager<User> UserManager
 
 <ul class="navbar-nav align-bottom">
-@if (SignInManager.IsSignedIn(User))
-{
-    <li class="nav-item">
-        <a  class="nav-link text-dark" asp-area="Identity" asp-page="/Account/Manage/Index" title="Manage">Мій аккаунт</a>
-        @*@if (@UserManager.GetUserAsync(User).Result.Avatar !=null)
+    @if (SignInManager.IsSignedIn(User))
+    {
+        <li class="nav-item">
+            <a class="nav-link text-dark" asp-area="Identity" asp-page="/Account/Manage/Index" title="Manage">Мій аккаунт</a>
+            @*@if (@UserManager.GetUserAsync(User).Result.Avatar !=null)
         {
             <img style='width:80px; height:60px;' src="data:image/jpeg;base64,@Convert.ToBase64String(@UserManager.GetUserAsync(User).Result.Avatar)" />
         }*@
-    </li>
-    <li class="nav-item">
-        <form class="form-inline" asp-area="Identity" asp-page="/Account/Logout" asp-route-returnUrl="@Url.Page("/", new { area = "" })" method="post" >
-            <button  type="submit" class="nav-link btn btn-link text-dark logout">Вийти</button>
-        </form>
-    </li>
-}
-else
-{
-    <li class="nav-item">
-        <a class="nav-link text-dark" asp-area="Identity" asp-page="/Account/Login">Ввійти</a>
-    </li>
-}
+        </li>
+        <li class="nav-item">
+            <form class="form-inline" asp-area="Identity" asp-page="/Account/Logout" asp-route-returnUrl="@Url.Page("/", new { area = "" })" method="post">
+                <button type="submit" class="nav-link btn btn-link text-dark logout">Вийти</button>
+            </form>
+        </li>
+    }
+    else
+    {
+        <li class="nav-item">
+            <a class="nav-link text-dark" asp-area="Identity" asp-page="/Account/Login">Ввійти</a>
+        </li>
+    }
 </ul>
